python-script/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
databricks-sdk==0.52.0
Deprecated==1.2.18
dnspython==2.7.0
docker==7.1.0
easyocr==1.7.2
email_validator==2.2.0
fastapi==0.115.0
fastapi-cli==0.0.7
filelock==3.18.0
Flask==3.1.0
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.1
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
lazy_loader==0.4
lxml==5.3.2
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mlflow==2.22.0
mlflow-skinny==2.22.0
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.5
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
orjson==3.10.18
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.3
pdf2image==1.17.0
pillow==11.2.1
protobuf==6.30.2
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.11.5
pydantic-extra-types==2.10.5
pydantic-settings==2.2.1
#pydantic_core==2.23.2
Pygments==2.19.1
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.2.3
PyPDF2==3.0.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docs==0.1.0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.4
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==80.3.1
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
starlette==0.38.6
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.30.6
waitress==3.0.2
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
nltk==3.8.1
sentencepiece==0.2.0
pytesseract
python-dotenv
mistralai
PyMuPDF
